Cabalized the project.
This is a pretty cool project but it really was missing being cabal compatible so I wrote
this in and it now works. Enjoy.

I also updated the readme file so that you know how to use this. It should be really easy
for anybody to get this project to compile now just by using cabal-dev.